Changed how machine borders were added and called
Before, an array held only the machine borders being shown on screen. Every time a button was hit, screenInner was set blank and the entries being shown on the current page were added then shown.

Now, at the start of the program, all machine borders are added to an array and only those needed are called.